test: access control tests for db call caching
