+++ packages/payload/src/index.ts
@@ -1740,6 +1740,8 @@ export { formatErrors } from './utilities/formatErrors.js'
 export { formatLabels, formatNames, toWords } from './utilities/formatLabels.js'
 export { getBlockSelect } from './utilities/getBlockSelect.js'
 export { getCollectionIDFieldTypes } from './utilities/getCollectionIDFieldTypes.js'
+export { getEntityPermissions } from './utilities/getEntityPermissions/getEntityPermissions.js'
+export { getEntityPolicies } from './utilities/getEntityPolicies.js'
 export { getFieldByPath } from './utilities/getFieldByPath.js'
 export { getObjectDotNotation } from './utilities/getObjectDotNotation.js'
 export { getRequestLanguage } from './utilities/getRequestLanguage.js'
