+++ packages/payload/src/utilities/getEntityPermissions/getEntityPermissions.ts
@@ -80,6 +80,8 @@ const topLevelGlobalPermissions = ['read', 'readVersions', 'update']
  * We cannot include siblingData or blockData here, as we do not have siblingData available once we reach block or array
  * rows, as we're calculating schema permissions, which do not include individual rows.
  * For consistency, it's thus better to never include the siblingData and blockData
+ *
+ * @internal - this function may change or be removed in a minor release.
  */
 export async function getEntityPermissions<TEntityType extends 'collection' | 'global'>(
   args: Args<TEntityType>,
