+++ test/access-control/shared.ts
@@ -28,3 +28,6 @@ export const publicUserEmail = 'public-user@payloadcms.com'
 export const publicUsersSlug = 'public-users'
 
 export const authSlug = 'auth-collection'
+
+export const whereCacheSameSlug = 'where-cache-same'
+export const whereCacheUniqueSlug = 'where-cache-unique'
